feat: integrate turbo-cdn and restructure tests
- Replace reqwest with turbo-cdn for faster, more reliable downloads
- Remove OpenSSL dependency to resolve compatibility issues
- Move all tests from src modules to dedicated tests/ directory
- Update logging to show info level by default, debug with --verbose
- Add comprehensive test coverage for downloader, config, and commands
- Support archive extraction and multiple CDN sources via turbo-cdn

Resolves #45
