+++ crates/shimexe-core/Cargo.toml
@@ -21,11 +21,7 @@ dirs.workspace = true
 which.workspace = true
 tracing.workspace = true
 regex = "1.10"
-reqwest = { version = "0.12", features = [
-    "json",
-    "rustls-tls",
-    "stream",
-], default-features = false }
+turbo-cdn = { version = "0.2.0", features = ["rustls-tls"], default-features = false }
 tokio = { version = "1.0", features = ["full"] }
 async-trait = "0.1"
 futures-util = "0.3"
